requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.109.2
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
install==1.3.5
Markdown==3.5.2
openai==1.12.0
pdfkit==1.0.0
pillow==10.2.0
pydantic==2.6.1
pydantic_core==2.16.2
PyMuPDF==1.23.22
PyMuPDFb==1.23.22
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.31.0
sniffio==1.3.0
starlette==0.36.3
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.1
zipp==3.17.0
